(possibly?) fixed a fact about eigenvectors
Was this meant to say what I've proposed? Because as stated, I am not sure if the proposition is correct. Let for a counter example A = [0 1; 1 0]. Then consider the column vectors a=[1; 1] and b=[-1; 1]. For the eigenvalue y=1, both vectors a and b and eigen vectors (because A*v = 1*v for both a and b), correct? So I think one eigenvalue may have many associated eigenvectors; but one eigenvector (I believe) always a unique eigenvalue.